Fix: Help text for encrypted volume type has incorrect description
Volume type encryption help text states that "The Key Size is the
size of the encryption key, in bits (e.g., 128, 256).
If the field is left empty, the provider default will be used."
This is not really correct, because 128 bit keys do not actually
work with LUKS as used by Cinder volumes.
It should only recommend 256 bit keys.
This patch fixes the issue.

Change-Id: I038918e4beef2ea4d346127713855e0421e8b721
Closes-Bug: #1780457