Drop unnecessary LOCALE_PATHS from settings.py
Django looks for and uses if it exists a locale directory in each of
the installed apps listed in INSTALLED_APPS [1]. horizon and
openstack_dashboard are specified in INSTALLED_APPS, so there is no need
to specify horizon/locale and openstack_dashboard/locale explicitly.
We can drop them without any side-effect.

[1] https://docs.djangoproject.com/en/1.11/topics/i18n/translation/#javascript-catalog-view

Change-Id: Ibe364b2a16894a51fe17411f29a326aef621472c
Related-Bug: #1804289