Instance: show network info even if name is empty
Previously, when a network name is an empty string,
network information of server interfaces will be empty,
so we cannot distinguish server interfaces.

Also network name in the instance detail page is capitalized
but there is no need to capitalize it and such information like
network name should be displayed as-is. This is not directly
related to the bug fixed by this commit, but it is simple enough
and related to network name, so I believe it is okay.

Change-Id: I2f2ad9b6bf50a2b53e79e4929d877111f875bd5b
Closes-Bug: #1805812