Allow to run pep8/docs tox env with horizon plugins
As of now, when horizon plugins are configured in the horizon repo,
pep8 and docs tox env cannot run.
This is because openstack_dashboard.settings tries to load horizon plugins
but horizon plugins are not installed in tox env by default and
the jobs will fail as a result.
This commit changes to use openstack_dashboard.test.settings
as DJANGO_SETTINGS_MODULE for these tox environments by default.

Change-Id: I6a9d0c74271cc4addd8a0964b5955d7d3b6d8267