Do not instantiate tables when a tab is not allowed
This commit chnages TableTab class not to instantiate
tables when the tab is not allowed.

Previously table(s) in a tab is always instatiated
regardless of that the tab is allowed or not.

We usually hide a tab using "allowed" method or "permissions" attribute
of the tab. Table(s) in the tab usually assumes that its dependencies
are satisfied and they are already checked by allowed or permissions.

Change-Id: I51d92221fdf02d836e1dc9343e85a69109159521
Closes-Bug: #1791296