Do not load openstack_dashboard.settings in unit tests
The previous behavior depends on different values for
SELECTABLE_THEMES and it seems this is the reason that test_themes
imported openstack_dashboard.settings.
If we override a value in django.conf.settings, we can use
override_settings decorator. This commit uses it and
we can now avoid importing openstack_dashboard.settings in UT.

Change-Id: Id1b7b7e431b86e725ad897613ca38f423a67e52d
Closes-Bug: #1809842