Fix osprofiler support in horizon
This patch is fixing usage of osprofiler in horizon.
It has been merged once, but after merge to master,
manage.py compress starts to fail with python2.7.

This fail was caused by "TypeError: 'encoding' is an
invalid keyword argument for this function" which is
now fixed by using six library.

Change-Id: I21aa8d49833bef9e7673958d15674faec82a0925
Closes-Bug: #1814028
Related-Bug: #1811261