Correct error message for volume size in Launch Instance
In the "Source" tab of the "Launch Instance" form,
when some image is selected as a boot source and
"Create New Volume" is selected, if a decimal number
such as 1.1 is provided in the volume size, the error message
still says "The volume size must be at least 1 GB" even if
the specified value is larger than min_disk.
This is incorrect. This commit corrects the message
for this case.

Change-Id: I36d37407eb208586fb7bacdb941b445c5ea33bcb
Closes-Bug: #1812959