Ensure django.contrib.auth.middleware monkey patching for APIMockTestCase
This patch is directly proposed to stable/rocky.

A cherry-pick of I6b2be31568e4aa4273990f9cfa76ca139620c84b
to stable/rocky is insufficient. The cherry-pick forgot to
handle APIMockTestCase as the corresponding code is different
between the master branch and stable/rocky branches and
as a result APIMockTestCase in stable/rocky branch does not
monkey-patch django.contrib.auth.middleware.

Related-Bug: #1806053
Change-Id: I4993ff5006b6366df64b6d96acddf15be0cc331c