Quote cinder qos-spec key name
In "Manage Specs" table of "Volume Types" panel, when a key name of
a QoS spec contains multiple spaces the QoS spec cannot be deleted
because multiple spaces in the key name are reduced into a single
space by the "minifyspace" templatetag.
"minifyspace" itself is useful, so it looks better to escape a key
name of a QoS spec to keep multiple spaces. This commit escapes
a key name of a QoS spec by using urllib.parse.quote().

Change-Id: Id11080e2a4284d69b8afb5c554728a98e024289f
Closes-Bug: #1811644