The id and track id use different value
The row data could be updated depend on track id,
and id could be as the identity of the row.
The table row checkbox bind to the model value by
the id.
Track id is "image_id"_"updated_time" of the associated
case, and it will be changed when the row updated.
So, the check status of the checkbox will be lost for
id equals track id. And the batch delete button also
could delete the images selected before.
This commit difference the id and track id, and id will
not be changed when update.

Change-Id: Icd3ad4c43e389f2d65389a53351d6f9ef666a341
Closes-Bug: #1816531