Adds community image loading for instance index view
This patch adds community image loading to the Instances index view
so that image names may be displayed in this view even for community
images.

Note, that a second image list API call had to be added to achieve
this, because the glance Rest API does not currently support adding
multiple visiblity values to it's GET filters and community images
have to be retrieved explicitly.

Change-Id: I55249de8762e0744b69ff655fcd7a4aeb56ba9d6
Closes-Bug: #1818823