network topology: handle port AZ correctly
Previously d3 version of the network topology view handles ports with
device_owner 'compute:nova' and 'compute:None' specially. This leads
to the situtaion that neutron ports with non-default AZ are not shown
properly. There is no reason to handle neutron ports with the default
AZ differently. What we would like to do is just to classify neutron
ports attached to nova servers.

This commit changes the logic to check device_owner of ports have
a prefix 'compute:'. Note that we also need to check if device_owner
is a string before checking the prefix.

Change-Id: I472bae9af3d14d8a31efcae8a7610b84c4f09c3d
Closes-Bug: #1820260