Do not call handle() when AZ is required
On "Edit Host Aggregation" form under Host Aggregates panel,
if the availability zone is cleared, both the error message and
the success message will appear.

In the aggregate API, nova does not allow to clear AZ once
it is set. This commit moves the check for availability_zone field
to the form definition. handle() method is now never called for
such case and the error message will be shown around the field.

Co-Authored-By: Akihiro Motoki <amotoki@gmail.com>
Change-Id: I74e03a47cee6915e9c24c6cf5c3264bf8b0e39b8
Closes-Bug: #1822971