Correctly set content-type for workflow response
The default content-type is html, and under certain
circumstances like devstack the html has some additional
comments added. Because this specific content is meant
to be parsed as json those extra html comments break the
modal.js code.

We can't set the content-type to json without rewriting parts of
modal.js, as it will will parse it into an object early and then
modal.js will fail as it attempts to parse an object into json.

Change-Id: If5a9ee7fe4273dc368dec7767a55d4467bf6a1a7
Closes-bug: #1824484