Do not try to access sets by index
code in neutron api in some circumstances tries to access
a set or a frozenset by index, which does not work.
In particular it may manifest as

Unable to connect to Neutron: 'frozenset' object has
no attribute '__getitem__'

error in the horizon.log when opening an empty instances list page.

Convert to list instead, and skip conversion for `collections.Sequence`
objects only that support all the methods used for this object
further in this method.

Closes-bug: #1825961
Change-Id: I141a28d96f71c06a1ebe44d7067ccf4609e22db6