Add relnote on cinderclient minver bump to 4.0.1
A fix for bug 1824017 bumped the minimum version requirement of
python-cinderclient to 4.0.0. The fix was backported to stable/stein,
so the minimum version bump should be covered by a release note.

python-cinderclient requirement is bumped to >=4.0.1
as 4.0.0 is blocked by global-requirements.txt [1].

It was suggested by the release team in the review of
horizon stein update release [2].

[1] https://opendev.org/openstack/requirements/src/commit/1baf07a77333fc57d9d4a669d0265b0e687fd031/global-requirements.txt#L256
[2] https://review.opendev.org/#/c/655447/

Change-Id: I2b43e159da7e58980b2810a104ae2418f237d1c7
Related-Bug: #1824017