Update URLs of horizon plugin registry
As result of the recent opendev migration, github repositories
for non-TC-governed OpenStack projects are not synced automatically.
group-based-policy-ui is such repository.
It is a good chance to point repositories on opendev.org.
This commit replaces github links to opendev links.

This commit also updates bug tracker links to more direct links
to the bug trackers.

sphinx extlink is now used to simplify the plugin registry doc.

Change-Id: Ib1a746e2cd6c7ed0c81c10e2767d6f76ad0203b9