Make project volume group table work even with volumev3 only env
Previously the volume group table in the project dashboard checked if
'volume' endpoint is available and 'volumev3' endpoint was not checked.
'volumev3' is an endpoint name commonly used for Volume v3 API and
it is not surprising that only volumev3 is configured in deployments,
so we should ensure that horizon works with environments only with
'volumev3' endpoint.

The service endpoints are already checked in panel.py, so there is no
need to have the permission check (i.e., service check) in the table.

Similar patterns are found in tables in project/snapshots and
project/vg_snapshots. They are also cleanup by this commit.

Change-Id: I12d03745bec5d8dec10a8e2b4306c595f2f39560
Closes-Bug: #1829461