Manually registering 2 Dashboards breaks get_dashboards() (python3)
The get_dashboards() method used sorted(<values>) to sort the
dashboards that have been registered manually.  This breaks if
two or more are registered that way because the Dashboard class
doesn't define a natural ordering.  The fix is to sort using
the Dashboard name as the sort key.  (Unit test included.)

This problem doesn't occur in python 2.7.

P.S.
It turns out that AdminPanel in horizon/test/unit/test_base.py is not
used during the review. It is dropped too as it is a simple change.

Change-Id: I6be10a92a1f43b3e0effee8148987c7e2c5690ff
Closes-bug: #1836295