Fix boot-from-volume test in project.instances.tests
The test for the instance table with boot-from-volume had
several issues previously.

a) api.cinder.volume_list is mocked but return_value was not
   specified and called arguments were not checked.
b) The mocked api.nova.server_list returned Server instance
   from novaclient, but it should return Server instance from
   the horizon API wrapper (api.nova.Server). As a result,
   for example, "image_name" property in api.nova.Server
   was skipped.
c) Test data for cinder volumes had wrong attachment information.
   attachment ID must match a volume ID, but it did not.

This commit fixes the above issues and adds a test logic to
check volume_image_metadata handling in instances/views.py.

This is a preparation to implement a unit test in
https://review.opendev.org/#/c/668595/.

NOTE:
The issues (a) and (b) above need to be fixed globally
in the openstack_dashboard testing, but it will be covered
by follow-up commit(s). This commit focues to prepare for
https://review.opendev.org/#/c/668595/.

Change-Id: I7ebb68220f75fef36b43b89e32e5aff18fc0fe25