Preserve form field order in python 3
In python 3, the ExternalUploadMeta metaclass does not preserve the
order of the fields of the class it is modifying.
First we add a __prepare__() method which returns an empty
collections.OrderedDict.
That mapping records the methods and attributes of the class it is
metaclassing as they are defined within the body of the class statement.
Then we send a collections.OrderedDict to create the new class.

Change-Id: I7223ae5bd931202c95602ec9674ae86b4708368d
Closes-Bug: #1835724