Display password_regex_description when password change fails
When keystone has password_regex set and the new password doesn't match
it, the attempt to change the password will fail, and keystone will
helpfully include password_regex_description in its error message.
Unfortunately Horizon discards that and displays "Unable to update the
user password.", which is considerably less helpful.

This change extracts the description from the error message in this
case, and displays it to the user together with a translated error
message. Unfortunately we have no way of also translating the
description.

Closes-bug: #1838003
Change-Id: I535c020deed76188c0f7150072fbdced9c5454fc