Set log devel to DEBUG in local_settings.py.example
Log level 'DEBUG' is useful to debug our CI.
local_settings.py.example is used as a template of local_settings.py
in our CI and it is not easy to change the log level in
local_settings.py using a shell script.
This commit sets the log level of 'console' to 'DEBUG'
when DEBUG settings is true.

Note that this does not mean the change of the default value.

Change-Id: If767cbc6c027060d564bdc2839bf6e61f6425d01