Add Allowed Address Pair/Delete buttons are only visible to admin
Initially In Queens, when logging into Horizon in context
of a _member_ in any project, the "Add Allowed Address Pair"
and "Delete" buttons in "Network" -> "Networks" ->
<select a network> -> "Ports" -> <select a port> ->
"Allowed Address Pairs" are not visible.

When accessing the same panel in context of a project where the user
has the admin role, the "Add" and "Delete" button is visible and
functional.

This patch fixes the above issue by updating the policy_target.

Change-Id: Ia7db6c2881c72580a156eb97b11bb51f295cfbce
Closes-Bug: #1794421
(cherry picked from commit 87b57dfe7113da09715d5613ab9c91c15a75c189)