django22: The 'login' functional-based view is gone (pt. 2)
This is the larger change. We could attempt to rewrite this whole thing
to a class-based view but that's more work that it's worth right now.
Instead, we simply do what the now-removed 'login' function-based view
was doing, as seen at [1].

A lot of never-used customizability is removed because it's not needed
and made things more complicated than it needed to be.

[1] https://github.com/django/django/blob/1.11/django/contrib/auth/views.py#L133-L139

Change-Id: Ib934d8a2c32cb32761558a68f061f415bb8737c4
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>