Handle deprecation of inspect.getargspec
getargspec has been deprecated in py3 with plans to remove it in py3.6.
The recommendation is to move to inspect.signature, but the results of
that call are different than the existing one.

There is also getfullargspec available under py3 that was originally
deprecated, but for the sake of handling 2/3 code, it has been
un-deprecated. This call uses inspect internally, but returns a mostly
compatible result with what getargspec did. This handles getargspec
deprecation by just using getfullargspec instead if it is available

Change-Id: Id0104f7e2b4c43725cb6bcd5e0cd905b19e3133e
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>