Updated max-width to be dynamic for .member class
When member name is longer than $members-list-item-width
(130px) in members list (for example, project group in Edit
Project), current text-overflow will put ... (ellipsis) at
the end of the name and name will be truncated.

Updated the styling of .member class to make max-width
dynamic 80% and wrap the name in case it is really long.

Change-Id: Ic295a9b1e7fd3525c633d51a60054e713da6a657
Closes-bug: #1843480