Don't display expiration warning for expiration date in the past
If the expiration date of a password is in the past, don't display the
expiration warning, as it's not helpful and confuses the user.

This can happen after a user whose password was set to be changed on the
first login changes it, but probably also in other situations, depending
on how many more bugs in Keystone there are around expiring passwords.

Change-Id: Ib79f6ef354c456bbdc2d7c1d4371ae15e825b557