Relnote: default values are moved to openstack_dashboard/defaults.py
In Train release, we move most settings from local_settings.py.example
to openstack_dashboard/defaults.py. It is worth covered in the Train
release notes so that operators who depend on local_settings.py.example
can be aware of the change.

Also drops settings related to the operation log feature in
local_settings.py.example are as they are defined in
horizon/defaults.py (which is loaded by
openstack_dashboard/defaults.py).

Change-Id: Ib06925aef305ffa1545d10529acd42d04177f5e1