Avoid duplicated maintenance of integration test config
Configurations of the integration tests are maintained both
in horizon.conf and in-code configuration. It is redundant.

Configurations of the integration tests are defined by oslo.config.
We can use default values from them and generate a sample config file
using oslo-config-generator.

This commit drops horizon.conf and generates the sample config.
We no longer have horizon.conf, so the logic to load the config files
is updated accordingly.

Also makes the usage of single and double quotes in config.py consistent.

Change-Id: Id16c9fd4114316ade63c74097e486623d0c5be9f