Set up the initial displayed values in all boot-sources
Problem description
===================

In the launch instance workflow while creating a virtual machine,  when
I select the Source tab, most of the time the default boot-source is
`Image` and everything works just fine. However, some times, the default
boot-source comes as Volume, due to some race condition (tested and
reproduced in Chrome 77). In this case, the Available items table is
empty, even if there are bootable volumes to be attached to the VM. This
problem happens because the initial state of boot-sources is setting
only the `displayedAvailable` of the Image source (that is the "default
source"), instead of setting all the initial `displayedAvailable`
elements.

Proposal
========

The proposal is basically to set up the `displayedAvailable` in
all boot-sources.

Closes-Bug: #1848334

Change-Id: I1c9399bf3f6c859206c093452c2e637892aa9722