Assume features from N-4 release instead of N-3
Curretnly, features from N-3 release are assumed as a baseline
in the horizon development policy document.
LTS of Ubuntu Cloud Archive supports upgrade from N-4 to N.
If we assume a case where horizon is upgraded first, N release of
horizon needs to be able to run with N-4 release of back-end services.
The difference of N-3 and N-4 would be small, so it sounds reasonable
to assume features from N-4 release.

N-4 is chosen considering LTS support in the major OpenStack distributions:
* Red Hat OpenStack Platform (RHOSP): X -> X+3
  https://access.redhat.com/support/policy/updates/openstack/platform/
* SuSE OpenStack Cloud (SOC): X -> X+2
  https://www.suse.com/releasenotes/x86_64/SUSE-OPENSTACK-CLOUD/8/#Upgrade
* Ubuntu Cloud Archive: X -> X+4
  https://www.ubuntu.com/about/release-cycle

Change-Id: If7007e9d5b912b48d6bf3e4fac5decc53cd8388b