Complete "Drop Python 2.7 Support" goal
The goal definition in the TC repo [1] defines the completion criteria.
We have already completed 1, 2, 3 and partially 4 in [1].
This commit completes 4 (requirements.txt etc) and 5 (release note).

Python dependency in the installation guide is updated.
Django support plan in Ussuri is updated as well.
References to python 2.7 in the document are updated as well.

[1] https://governance.openstack.org/tc/goals/selected/ussuri/drop-py27.html

Change-Id: I0383920e32e053e1611519be774fab8dec124ccc