Make tox -e runserver work faster
Compiling translation messages on every run of runserver makes it very
slow and inconvenient for development. Editing of the translations is
done rarely, and compilemessages can be run manually when it does
happen.

I'm also adding -v 0 to all the manage commands, to make it output
less garbage to the screen, making it a little easier to find actual
errors during development.

Change-Id: I95db4505c3a9cd5b2ecc3e6ad6a30d438a4447b9