GUI hangs on creation of project in User creation form
On creation of project in user creation form, the response is sent
as html and the javascript try to parse the output as json and fails.

The project creation response should be sent as text/plain so that
javascript can parse the string as json.

The bug have 3 parts
1. Users are not listed on GUI in different domain
2. Projects are not listed on GUI in different domain
3. Gui hangs during creation of user/project using + option
Case 1 & 2 are fixed at https://review.opendev.org/#/c/697962/
This patch fixes case 3.

Change-Id: I4936a158ad99cf9d7514e79c06ed2315ce11218b
Closes-Bug: #1826114