Cap hacking version to <2
Hacking 2.0.0 and its dependencies like flake8 have incompatible changes
and bumping to them leads to pep8 check failure.
It is good to address pep8 warnings in the master branch, but in case of
stable branches it is good to continue to use the same version of hacking.
This commit is proposed to the master branch so that it can be backported.

Change-Id: I2b174ef3231956037650eba19aebe5989dffa617