Deprecate OPENSTACK_NOVA_EXTENSIONS_BLACKLIST
All of the nova API extensions have been mainlined several releases ago
and there is no potential performance issue in the nova API.
This option is used only to toggle features in horizon and
there seems no performance issues controlled by the option in horizon.
Considering this situation, this option is deprecated now.

Change-Id: If8e58e8324bad919ea45d0ccb10e21371d97f470