Generate plugin registry based on releases repo
The plugin registry is maintained manually, so it tends to be out-of-date.
The deliverble files in openstack/releases repository provide information
on repositories and bug trackers. It sounds reasonable to me to generate
the plugin registry based on infomration in the openstack/releases repo.

The script added picks up deliverables marked as horizon-plugin.
mistral-dashboard and networking-bgpvpn dashboard are released in
different ways, so they are handled specially in the script.

How to run:
  ./tools/list-horizon-plugins.py \
      --csv-file plugin-registry.csv \
      ../releases ussuri

Note that this commit does not autogenerate the plugin registry.
The script still needs to be run manually.
Further automation would be the next step.

Note that bug trackers of senlin-dashboard and trove-dashboard in
the releases repo looks out-of-date. I will submit a fix for them soon.

The downside of the automation is that we cannot list non-official
horizon plugins but I believe it is a good compromise.
The only non-offical plugin listed now is the group-based-policy UI,
but it is questionable whether it is still maintained well.

Change-Id: Id56be13fb93c6dc123e7cfb1c906357cdf59732b