zuul: Avoid duplicated job
Currently openstack-tox-py36/37 jobs use Django 2.2.
horizon-non-primary-django-jobs template just runs Django 2.2 test.
These two are duplicated.

To keep the zuul project template (as it is used in horizon plugins),
a job in openstack-python3-ussuri-jobs is specified in the template.
Zuul allows to specify a job in multiple places, so it should work.

Change-Id: I248ac27b60836ce0da473f452ce017542139c671