Use python 2.7 as the default interpreter in tox
Infra uses Python 3.5 of tox jobs.
The pep8 job and py27djNNN jobs are now run with python3.5
while they are expected to run with python2.7 as of the release.

The pep8 job fails on Python2-specific code in the
openstack_dashboard/dashboards/admin/images/views.py module with 'long'
type cast. This patch enforces us to use Python 2.7 for pep8 job.

This probably could be fixed in the infra side but as a short term
solution we can cope with this situation in the horizon repo side.

Co-Authored-By: Akihiro Motoki <amotoki@gmail.com>
Change-Id: I25b0b00eb3d575f8a8c1705f987cfcc0336eccfb